checkpoint: some G_CULLDL emulation progress
This is a checkpoint for current progress on G_CULLDL. It's imperfect, so two things to keep in mind for the time being:
1. It sometimes erroneously thinks culling volumes fall outside the camera's view, so geometry may sometimes disappear at certain camera angles.
2. It slows down when you move the camera around at first as it builds its texture cache. (This will be remedied later by doing an initial pass on-load with culling disabled.)
On the plus side, Fire Temple performance has jumped from ~80 FPS to anywhere between ~70 FPS to ~300 FPS (and beyond; averaging ~180 FPS) on my machine depending on which room you're in and how much off-screen geometry is able to be skipped during rendering. As the culling algorithm is improved, I expect these numbers to rise.